+++ src/etherRuntime.ts
@@ -5,16 +5,16 @@
 import { readFileSync } from 'fs';
 import { EventEmitter } from 'events';
 
-export interface MockBreakpoint {
+export interface etherBreakpoint {
 	id: number;
 	line: number;
 	verified: boolean;
 }
 
 /**
- * A Mock runtime with minimal debugger functionality.
+ * A ether runtime with minimal debugger functionality.
  */
-export class MockRuntime extends EventEmitter {
+export class etherRuntime extends EventEmitter {
 
 	// the initial (and one and only) file we are 'debugging'
 	private _sourceFile: string;
@@ -28,8 +28,8 @@ export class MockRuntime extends EventEmitter {
 	// This is the next line that will be 'executed'
 	private _currentLine = 0;
 
-	// maps from sourceFile to array of Mock breakpoints
-	private _breakPoints = new Map<string, MockBreakpoint[]>();
+	// maps from sourceFile to array of ether breakpoints
+	private _breakPoints = new Map<string, etherBreakpoint[]>();
 
 	// since we want to send breakpoint events, we will assign an id to every event
 	// so that the frontend can match events with breakpoints.
@@ -100,12 +100,12 @@ export class MockRuntime extends EventEmitter {
 	/*
 	 * Set breakpoint in file with given line.
 	 */
-	public setBreakPoint(path: string, line: number) : MockBreakpoint {
+	public setBreakPoint(path: string, line: number) : etherBreakpoint {
 
-		const bp = <MockBreakpoint> { verified: false, line, id: this._breakpointId++ };
+		const bp = <etherBreakpoint> { verified: false, line, id: this._breakpointId++ };
 		let bps = this._breakPoints.get(path);
 		if (!bps) {
-			bps = new Array<MockBreakpoint>();
+			bps = new Array<etherBreakpoint>();
 			this._breakPoints.set(path, bps);
 		}
 		bps.push(bp);
@@ -118,7 +118,7 @@ export class MockRuntime extends EventEmitter {
 	/*
 	 * Clear breakpoint in file with given line.
 	 */
-	public clearBreakPoint(path: string, line: number) : MockBreakpoint | undefined {
+	public clearBreakPoint(path: string, line: number) : etherBreakpoint | undefined {
 		let bps = this._breakPoints.get(path);
 		if (bps) {
 			const index = bps.findIndex(bp => bp.line === line);